+++ pr1/zavd14/rbuf.h
@@ -10,7 +10,7 @@ typedef struct RingBuffer {
 	size_t size;
 } RingBuffer;
 
-void rbuf_init(RingBuffer *rb, int n);
+void rbuf_init(RingBuffer *rb, size_t n);
 void rbuf_push(RingBuffer *rb, const char *line);
 void rbuf_print(RingBuffer *rb, bool reverse);
 #endif
